refactor(angular-components): move table data of vehicle basics tab to the vehicle basics tab component
* Move table data of vehicle basics tab to vehicle-view-basics-tab-content component also the setTableData method from vehicle-view component so vehicle-view component doesn't have to carry both logics and property of basics tab and photos tab so that vehicle-view component can be lighter and cleaner.
* I also add edit button which will navigate user to the edit form page of the specific vehicle he/she view currently in the vehicle basics tab
* Add delete button in the vehicle basics tab so user can delete the vehicle in this page.
