+++ ClientApp/src/app/vehicle-view/vehicle-view-basics-tab-content.component.ts
@@ -1,4 +1,7 @@
-import { Component, Input } from "@angular/core";
+import { Component, Input, OnInit } from "@angular/core";
+import { VehicleService } from "../vehicle.service";
+import { ActivatedRoute, Router } from "@angular/router";
+import { ToastrService } from "ngx-toastr";
 
 @Component({
   selector: 'app-vehicle-basics-tab-content',
@@ -26,9 +29,69 @@ import { Component, Input } from "@angular/core";
           </li>
         </ul>
       </div>
+      <div class="mt-2">
+        <a [routerLink]="['/vehicles', vehicleId]" class="btn btn-primary" role="button">Edit</a>
+        <button type="button" class="btn btn-danger ml-2" (click)="delete()">Delete</button>
+      </div>
     </div>
   `
 })
-export class VehicleViewBasicsTabContentComponent {
-  @Input('table-data') tableData;
+export class VehicleViewBasicsTabContentComponent implements OnInit{
+  tableData = [
+    { title: 'Make', key: 'make.name', value: '', type: 'string' },
+    { title: 'Model', key: 'model.name', value: '', type: 'string' },
+    { title: 'Contact Name', key: 'contact.name', value: '', type: 'string' },
+    { title: 'Contact Phone', key: 'contact.phone', value: '', type: 'string' },
+    { title: 'Contact Email', key: 'contact.email', value: '', type: 'string' },
+    { title: 'Registered', key: 'isRegistered', value: '', type: 'string' },
+    { title: 'Features', key: 'features', value: [], type: 'array' },
+  ];
+  vehicleId = 0
+
+  constructor(
+    private vehicleService: VehicleService,
+    private route: ActivatedRoute,
+    private router: Router,
+    private toastr: ToastrService
+  ) {
+    this.vehicleId = +this.route.snapshot.paramMap.get('id');
+  }
+
+  ngOnInit(): void {
+    this.setTableData();
+  }
+
+  private setTableData() {
+    this.vehicleService.getVehicle(this.vehicleId).subscribe(vehicle => {
+      this.tableData = this.tableData.map(row => {
+        if (row.type === 'string') {
+          // Get the property value from the object.
+          // eg: vehicle: { contact: { name: 'abc' } } so I need to get vehicle['contact]['name'] value.
+          const value = row.key.split('.').reduce((acc, key) => {
+            return acc[key];
+          }, vehicle);
+          return {
+            ...row,
+            value: value
+          };
+        } else if (row.type === 'array') {
+          // If row type is array, I will process it in the html. because it is easier to iterate it and display the
+          // elements inside the array.
+          return {
+            ...row,
+            value: vehicle[row.key]
+          };
+        }
+      });
+    });
+  }
+
+  delete() {
+    if (confirm('Are you sure?')) {
+      this.vehicleService.delete(this.vehicleId).subscribe(() => {
+        this.toastr.success('Vehicle has been deleted', 'Success');
+        this.router.navigate(['/']);
+      });
+    }
+  }
 }
