+++ ClientApp/src/app/vehicle-view/vehicle-view.component.html
@@ -2,11 +2,10 @@
 
 <div class="card">
   <app-card-tabs>
-    <!--    <app-tab [active]="true" tab-title="Basics">-->
-    <app-tab [active]="false" tab-title="Basics">
-      <app-vehicle-basics-tab-content [table-data]="tableData"></app-vehicle-basics-tab-content>
+    <app-tab [active]="true" tab-title="Basics">
+      <app-vehicle-basics-tab-content></app-vehicle-basics-tab-content>
     </app-tab>
-    <app-tab [active]="true" tab-title="Photos">
+    <app-tab tab-title="Photos">
       <div class="input-group mb-3">
         <div class="input-group-prepend">
           <span class="input-group-text" id="inputGroupFileAddon01">Upload</span>
